code cleanup

+++ bin/css64.js
@@ -11,6 +11,13 @@ var opts = {
 var theme = process.env.THEME || 'tkw';
 console.log("Converting background url to inline base64 with theme '" + theme + "'");
 
-src = fs.readFileSync('themes/' + theme + '/style.css');
+var filePath = 'themes/' + theme + '/style.css';
+
+if (!fs.existsSync(filePath)) {
+  console.error(filePath + " does not exist - required");
+  process.exit(1);
+}
+
+src = fs.readFileSync(filePath);
 output = postcss().use(base64(opts)).process(src).css;
 fs.writeFileSync('themes/' + theme + '/style.64.css', output);
